// bachelors-project/src/modules/utils/interfaces/register.dto.ts
import { Gender } from "../enums/gender.enum";

export interface RegisterDto {
    email:string;
    password:string;
    verifyPassword: string;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    address: string;
    city: string;
    countryCode: string;
    socialSecurityNumber: string;
    occupation: string;
    companyInfo: string;
    gender: Gender
}